+++ src/core/index.ts
@@ -1,35 +1,35 @@
 import { Utilities } from "@/interfaces";
 
-import { backgroundUtils } from "./background";
-import { borderUtils } from "./border";
-import { boxModelUtils } from "./box-model";
-import { colorUtils } from "./color";
-import { effectUtils } from "./effect";
-import { flexboxUtils } from "./flexbox";
-import { fontUtils } from "./font";
-import { gridUtils } from "./grid";
-import { interactivityUtils } from "./interactivity";
-import { outlineUtils } from "./outline";
-import { positioningUtils } from "./positioning";
-import { svgUtils } from "./svg";
-import { tableUtils } from "./table";
-import { textUtils } from "./text";
-import { transformUtils } from "./transform";
+import { background } from "./background";
+import { border } from "./border";
+import { boxModel } from "./box-model";
+import { color } from "./color";
+import { effect } from "./effect";
+import { flexbox } from "./flexbox";
+import { font } from "./font";
+import { grid } from "./grid";
+import { interactivity } from "./interactivity";
+import { outline } from "./outline";
+import { positioning } from "./positioning";
+import { svg } from "./svg";
+import { table } from "./table";
+import { text } from "./text";
+import { transform } from "./transform";
 
-export const allUtils: Utilities = {
-  ...backgroundUtils,
-  ...borderUtils,
-  ...boxModelUtils,
-  ...colorUtils,
-  ...effectUtils,
-  ...flexboxUtils,
-  ...fontUtils,
-  ...gridUtils,
-  ...interactivityUtils,
-  ...outlineUtils,
-  ...positioningUtils,
-  ...svgUtils,
-  ...tableUtils,
-  ...textUtils,
-  ...transformUtils,
+export const core: Utilities = {
+  ...background,
+  ...border,
+  ...boxModel,
+  ...color,
+  ...effect,
+  ...flexbox,
+  ...font,
+  ...grid,
+  ...interactivity,
+  ...outline,
+  ...positioning,
+  ...svg,
+  ...table,
+  ...text,
+  ...transform,
 };