+++ src/index.ts
@@ -1,38 +1,40 @@
-import { allUtils } from "./core";
-import { backgroundUtils } from "./core/background";
-import { borderUtils } from "./core/border";
-import { boxModelUtils } from "./core/box-model";
-import { colorUtils } from "./core/color";
-import { effectUtils } from "./core/effect";
-import { flexboxUtils } from "./core/flexbox";
-import { fontUtils } from "./core/font";
-import { gridUtils } from "./core/grid";
-import { interactivityUtils } from "./core/interactivity";
-import { outlineUtils } from "./core/outline";
-import { positioningUtils } from "./core/positioning";
-import { svgUtils } from "./core/svg";
-import { tableUtils } from "./core/table";
-import { textUtils } from "./core/text";
-import { transformUtils } from "./core/transform";
+import { core } from "./core";
+import { background } from "./core/background";
+import { border } from "./core/border";
+import { boxModel } from "./core/box-model";
+import { color } from "./core/color";
+import { effect } from "./core/effect";
+import { flexbox } from "./core/flexbox";
+import { font } from "./core/font";
+import { grid } from "./core/grid";
+import { interactivity } from "./core/interactivity";
+import { outline } from "./core/outline";
+import { positioning } from "./core/positioning";
+import { svg } from "./core/svg";
+import { table } from "./core/table";
+import { text } from "./core/text";
+import { transform } from "./core/transform";
 import { Utilities } from "./interfaces";
 
-// functions
-export const getAllUtils = (): Utilities => allUtils;
-export const getBackgroundUtils = (): Utilities => backgroundUtils;
-export const getBorderUtils = (): Utilities => borderUtils;
-export const getBoxModelUtils = (): Utilities => boxModelUtils;
-export const getColorUtils = (): Utilities => colorUtils;
-export const getEffectUtils = (): Utilities => effectUtils;
-export const getFlexboxUtils = (): Utilities => flexboxUtils;
-export const getFontUtils = (): Utilities => fontUtils;
-export const getGridUtils = (): Utilities => gridUtils;
-export const getInteractivityUtils = (): Utilities => interactivityUtils;
-export const getOutlineUtils = (): Utilities => outlineUtils;
-export const getPositioningUtils = (): Utilities => positioningUtils;
-export const getSvgUtils = (): Utilities => svgUtils;
-export const getTableUtils = (): Utilities => tableUtils;
-export const getTextUtils = (): Utilities => textUtils;
-export const getTransformUtils = (): Utilities => transformUtils;
+// all utilities
+export const coreUtils = (): Utilities => core;
+
+// individual utilities
+export const backgroundUtils = (): Utilities => background;
+export const borderUtils = (): Utilities => border;
+export const boxModelUtils = (): Utilities => boxModel;
+export const colorUtils = (): Utilities => color;
+export const effectUtils = (): Utilities => effect;
+export const flexboxUtils = (): Utilities => flexbox;
+export const fontUtils = (): Utilities => font;
+export const gridUtils = (): Utilities => grid;
+export const interactivityUtils = (): Utilities => interactivity;
+export const outlineUtils = (): Utilities => outline;
+export const positioningUtils = (): Utilities => positioning;
+export const svgUtils = (): Utilities => svg;
+export const tableUtils = (): Utilities => table;
+export const textUtils = (): Utilities => text;
+export const transformUtils = (): Utilities => transform;
 
 // types
 export type * from "./interfaces";